BUG: image.create_mask now returns ImageBase
The image.create_mask() function now returns an ImageBase rather than
a bare ndarray, so that it can be more easily saved with the transform
etc. In addition, the ImageBase class is modified to always copy the
transform during initialisation, to avoid accidentally overwriting the
transforms for parents during any modification.