BUG: Fixes for resample()
Previously sample had some issues where using a single slice for the
output shape would cause the slice dimension to be dropped when
transforming the coordinates which would interfere with doing the
interpolation. This is fixed by coercing the transformed coordinates
back into the target output shape.

In addition this brought to light another bug where the row_vector() and
col_vector() functions where swapped, which led to some transposement
problems.

Fixes #90
Fixes #92