BUG: Fixed dt parameter from SDAT
The value for dt in the Philips SDAT reading code was hardcoded.
This change fixes it to be read out of the file instead. There
seem to be multiple ways of accessing the dt parameter, none of
them that are simply called dt. Currently the value is computed
from the column extrema (start and end times of the FID) divided
by the number of columns.