BUG: Encoding issue with Philips SPAR files
Some SPAR files have been identified which do not match utf-8 encoding.
This is generally caused by patient names containing uncommon characters
such as accents. As it is currently unclear what the standard encoding
used by Philips is, and this may vary depending on system version or
location, an additional `spar_encoding` parameter is added to the
`load_sdat()` function where a different encoding can be specified if
necessary.