BUG: Loading DICOM files with non-Siemens extensions (#131)
BUG:  Loading DICOM files with non-Siemens extensions fails

This fix changes the way that load_dicom_volume() finds other DICOM
slices to assemble a volume. Previously, it used the extensions .ima
and .dcm to identify DICOM slices, however these are used by Siemens
and not GE or Philips scanners. Now the supplied slice extension is
used, all other files with the same extension in the same folder will
be opened (or an attempt will be made) and their SeriesInstanceUID
checked to see if they match.