ENH: Adding axis to channel combination functions
The functions to calculate channel combination weights,
`combine_channels()` and `svd_weighting()` previously required the
channels to be in the first axis of the data passed in to them. Now
we can instead pass an axis parameter into the function to specify
which axis contains the channels dimension. The default value of this
parameter is now -2, which is the default for Suspect loaded MRSData
objects.

Fixes #128