MAINT: v0.4.0 release commit
This commit makes a few minor changes to make the package compatible
with updates to upstream packages such as pydicom that have come out
of alpha status. It adds the missing dependency of parsley to the
requirements.txt and updates the supported Python versions, which are
now from Python 3.6 upwards. Finally it brings the version file up to
0.4.0