#BUG143 incorrect data shape in some Siemens .IMA files
This fix is to address the problem that for some Siemens .IMA files the
calculated data shape does not match the size of the accompanying data,
resulting in an Exception. The change means that in this situation the
data shape is adjusted to use the data size, rather than the (multi-
dimensional) shape calculated from the header info.

fixes #143