#BUG Sag voxels in twix fail to calculate a transform
This code addresses an issue reported in #145 for Siemens twix files
where voxels prescribed sagittally calculate a transform of NaNs. This
arises because in the twix file the row and column axes are not
specified, but must be calculated from the normal and a rotation from a
starting pair of axes. This starting pair of axes is not consistent but
depends on how the voxel was planned. In turn this is not described in
the file but must be deduced from the normal vector. The new code is an
attempt to solve this problem for the data files where an issue has been
reported, but may not be correct for some other class of twix files.

Fixes #145